requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
gunicorn==0.17.2
python-instagram==1.3.0